+++ test-suite/output/ZSyntax.v
@@ -19,7 +19,7 @@ Check (0 + Z.of_nat 11)%Z.
 
 (* Check hexadecimal printing *)
 Definition to_num_int n := Numeral.IntHex (Z.to_hex_int n).
-Numeral Notation Z Z.of_num_int to_num_int : Z_scope.
+Number Notation Z Z.of_num_int to_num_int : Z_scope.
 Check 42%Z.
 Check (-42)%Z.
 Check 0%Z.